# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Root endpoint"
          }
        },
        "summary": "Root endpoint",
        "tags": [
          "App"
        ]
      }
    },
    "/user/signup": {
      "post": {
        "operationId": "UserController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "User registration",
        "tags": [
          "User"
        ]
      }
    },
    "/user/signin": {
      "post": {
        "operationId": "UserController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "User login",
        "tags": [
          "User"
        ]
      }
    },
    "/user": {
      "get": {
        "operationId": "UserController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get current user profile",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update current user profile",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete current user account",
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get user by ID (Admin)",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update user by ID (Admin)",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_removeAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete user by ID (Admin)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/all": {
      "get": {
        "operationId": "UserController_getAllUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all users (Admin)",
        "tags": [
          "User"
        ]
      }
    },
    "/deck/review": {
      "post": {
        "operationId": "DeckController_submitReview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted successfully"
          }
        },
        "summary": "Submit card review",
        "tags": [
          "Deck"
        ]
      }
    },
    "/deck/review/{id}": {
      "get": {
        "operationId": "DeckController_getDueReviews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return due reviews"
          }
        },
        "summary": "Get due reviews for a deck",
        "tags": [
          "Deck"
        ]
      }
    },
    "/deck": {
      "post": {
        "operationId": "DeckController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeckDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deck created successfully"
          }
        },
        "summary": "Create a new deck",
        "tags": [
          "Deck"
        ]
      },
      "get": {
        "operationId": "DeckController_findAllCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all decks for current user",
        "tags": [
          "Deck"
        ]
      }
    },
    "/deck/by": {
      "get": {
        "operationId": "DeckController_findAllByUser",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all decks by user ID (Admin)",
        "tags": [
          "Deck"
        ]
      }
    },
    "/deck/{id}": {
      "get": {
        "operationId": "DeckController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a deck by id",
        "tags": [
          "Deck"
        ]
      },
      "patch": {
        "operationId": "DeckController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeckDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a deck",
        "tags": [
          "Deck"
        ]
      },
      "delete": {
        "operationId": "DeckController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a deck",
        "tags": [
          "Deck"
        ]
      }
    },
    "/card": {
      "post": {
        "operationId": "CardController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Card created successfully"
          }
        },
        "summary": "Create a new card",
        "tags": [
          "Card"
        ]
      },
      "get": {
        "operationId": "CardController_findAll",
        "parameters": [
          {
            "name": "deckId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all cards"
          }
        },
        "summary": "Get all cards",
        "tags": [
          "Card"
        ]
      }
    },
    "/card/{id}": {
      "get": {
        "operationId": "CardController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a card"
          }
        },
        "summary": "Get a card by id",
        "tags": [
          "Card"
        ]
      },
      "patch": {
        "operationId": "CardController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card updated successfully"
          }
        },
        "summary": "Update a card",
        "tags": [
          "Card"
        ]
      },
      "delete": {
        "operationId": "CardController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card deleted successfully"
          }
        },
        "summary": "Delete a card",
        "tags": [
          "Card"
        ]
      }
    }
  },
  "info": {
    "title": "FlashLearn",
    "description": "The FlashLearn API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "flashlearn",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "confirmPassword"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "id",
          "username",
          "email",
          "role",
          "createdAt",
          "lastLoginAt"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        }
      },
      "SubmitReviewItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "cardId": {
            "type": "number"
          },
          "quality": {
            "type": "string",
            "enum": [
              "Again",
              "Hard",
              "Good",
              "Easy"
            ]
          },
          "repetitions": {
            "type": "number"
          },
          "interval": {
            "type": "number"
          },
          "eFactor": {
            "type": "number"
          },
          "nextReviewDate": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "cardId",
          "quality",
          "repetitions",
          "interval",
          "eFactor",
          "nextReviewDate",
          "reviewedAt"
        ]
      },
      "SubmitReviewDto": {
        "type": "object",
        "properties": {
          "CardReviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmitReviewItemDto"
            }
          }
        },
        "required": [
          "CardReviews"
        ]
      },
      "CreateDeckDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateDeckDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateCardDto": {
        "type": "object",
        "properties": {
          "deckId": {
            "type": "number"
          },
          "front": {
            "type": "string"
          },
          "back": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          }
        },
        "required": [
          "deckId",
          "front",
          "back",
          "tags"
        ]
      },
      "UpdateCardDto": {
        "type": "object",
        "properties": {
          "front": {
            "type": "string"
          },
          "back": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          }
        }
      }
    }
  }
}